+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/nasekinid/Desktop/react/react-apps/srar-db/src/index.js":"1"},{"size":1559,"mtime":1610220271149,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"zc6d9b","/Users/nasekinid/Desktop/react/react-apps/srar-db/src/index.js",["6"],{"ruleId":"7","severity":1,"message":"8","line":9,"column":48,"nodeType":"9","messageId":"10","endLine":9,"endColumn":49},"no-useless-concat","Unexpected string concatenation of literals.","BinaryExpression","unexpectedConcat"]
+[{"/Users/nasekinid/Desktop/react/react-apps/srar-db/src/index.js":"1","/Users/nasekinid/Desktop/react/react-apps/srar-db/src/components/app/app.js":"2","/Users/nasekinid/Desktop/react/react-apps/srar-db/src/components/app/index.js":"3"},{"size":158,"mtime":1610221200488,"results":"4","hashOfConfig":"5"},{"size":229,"mtime":1610221110675,"results":"6","hashOfConfig":"5"},{"size":45,"mtime":1610221394002,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"zc6d9b",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/nasekinid/Desktop/react/react-apps/srar-db/src/index.js",[],"/Users/nasekinid/Desktop/react/react-apps/srar-db/src/components/app/app.js",[],"/Users/nasekinid/Desktop/react/react-apps/srar-db/src/components/app/index.js",[]]
