+++ src/components/app/app.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+import './app.css';
+
+export default class App extends Component {
+    render() {
+
+        return (
+            <div>
+                <h1>Hello!</h1>
+            </div>
+        );
+    }
+}
