+++ src/components/item-list/index.js
@@ -1,0 +1,3 @@
+import ItemList from './item-list';
+
+export default ItemList;
