+++ src/components/person-details/index.js
@@ -1,0 +1,3 @@
+import PersonDetails from './person-details';
+
+export default PersonDetails;
