+++ src/components/planet-details/index.js
@@ -1,0 +1,3 @@
+import PlanetDeatails from './planet-details';
+
+export default PlanetDeatails;
