+++ src/components/random-planet/index.js
@@ -1,0 +1,3 @@
+import RandomPlanet from './random-planet';
+
+export default RandomPlanet;
