+++ src/components/starship-details/index.js
@@ -1,0 +1,3 @@
+import StarchipDetails from './starship-details';
+
+export default StarchipDetails;
