+++ src/index.js
@@ -1,73 +1,6 @@
-class SwapiService {
+import React from 'react';
+import ReactDOM from 'react-dom';
 
-  _apiBase = 'https://swapi.dev/api';
+import App from './components/app';
 
-  async getResource(url) {
-    const res = await fetch(`${this._apiBase}${url}`);
-
-    if (!res.ok) {
-      throw new Error(`Could not fetch ${url}` + `, received ${res.status}`)
-    }
-
-    return await res.json();
-  }
-
-  async getAllPeople() {
-    const res = await this.getResource(`/people/`);
-    return res.results;
-  }
-
-  getPerson(id) {
-    return this.getResource(`/people/${id}/`);
-  }
-
-  async getAllPlanets() {
-    const res = await this.getResource(`/planets/`);
-    return res.results;
-  }
-
-  getPlanet(id) {
-    return this.getResource(`/planet/${id}/`);
-  }
-
-  async getAllStarships() {
-    const res = await this.getResource(`/starships/`);
-    return res.results;
-  }
-
-  getStarship(id) {
-    return this.getResource(`/starships/${id}/`);
-  }
-}//асинхронный источник данных
-
-const swapi = new SwapiService();
-
-// swapi.getAllPeople().then((people) => {
-//   people.forEach(p => {
-//     console.log(p.name);
-//   });
-// });
-
-swapi.getAllStarships().then((starship) => {
-  starship.forEach(s => {
-    console.log(s.name);
-  })
-})
-
-// const getResource = async (url) => {
-//   const res = await fetch(url);
-//   if (!res.ok) {
-//     throw new Error(`Could not fetch ${url}` + `, received ${res.status}`)
-//   }
-//   const body = await res.json();
-//   return body;
-// }
-
-// getResource('https://swapi.dev/api/people/1/')
-//   .then((body) => {
-//     console.log(body);
-//   })
-//   .catch((err) => {
-//     console.error('Could not fetch', err);
-//   });
-
+ReactDOM.render( <App/> , document.getElementById('root'));
